% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_data.R
\name{import_sdm_data}
\alias{import_sdm_data}
\title{Import data}
\usage{
import_sdm_data(filename, obs_coln, var_coln, na.rm = FALSE)
}
\arguments{
\item{filename}{A character string representing a path to a csv file.}

\item{obs_coln}{The name of the column with the species observations.}

\item{var_coln}{A vector with column names that will be used as predictors.}

\item{na.rm}{A logical value.}
}
\value{
A list with the elements 'data', 'subset', ...
}
\description{
Import data
}
\seealso{
\code{\link{read.csv}}
}

